Api Order and Stripe Payment (#20)

+++ LearnCSharp.Domain/Entities/OrderDetails.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int Quantity { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
         public int ProductId { get; set; }
         public Product Product { get; set; }
         public int OrderId { get; set; }
@@ -15,6 +15,6 @@
         public string SizeName { get; set; }
         public string CrustName { get; set; }
         public string Note { get; set; }
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
     }
 }
